Updated docker-entrypoint.sh and removed global journald.conf
Updated docker-entrypoint.sh and removed global journald.conf:
* docker-entrypoint.sh now only redirects journal to stdout when NOT launched with CMD
* removed the global journald.conf file in support of the above.

TODO: Review and consider the various needs for shipping journal logs to stdout of the TTY - it seems to me that there is an argument both for, and against it. Currently falling on the side of NOT shipping those logs when running with a CMD, as this image is designed to launch from the entrypoint, and thus a CMD generally means it was launched with a CMD from a `docker run` command-line execution.